Moved mouse tool tip to separate file.
