+++ src/vs15/CoCo/CoCo.csproj
@@ -49,6 +49,7 @@
       <Link>Properties\AssemblyInfoCommon.cs</Link>
     </Compile>
     <Compile Include="Guids.cs" />
+    <Compile Include="MouseTrackToolTipPresenter.cs" />
     <Compile Include="Services\AnalyzingService.cs" />
     <Compile Include="Migrations\MigrationService_2_3_0.cs" />
     <Compile Include="ClassificationManager.cs" />
@@ -164,6 +165,7 @@
     <Reference Include="System.ValueTuple, Version=4.0.1.1, Culture=neutral, PublicKeyToken=cc7b13ffcd2ddd51, processorArchitecture=MSIL">
       <HintPath>..\..\packages\System.ValueTuple.4.3.1\lib\netstandard1.0\System.ValueTuple.dll</HintPath>
     </Reference>
+    <Reference Include="System.Xaml" />
     <Reference Include="System.Xml" />
     <Reference Include="WindowsBase" />
   </ItemGroup>
