+++ src/vs15/CoCo/ToolTipPresenterFactory.cs
@@ -14,8 +14,6 @@ namespace CoCo
 
         public IToolTipPresenter Create(ITextView textView, ToolTipParameters parameters)
         {
-            return new ToolTipPresenter(viewElementFactoryService, textView, parameters);
-
             return parameters.TrackMouse
                 ? new MouseTrackToolTipPresenter(_viewElementFactoryService, textView, parameters)
                 : new ToolTipPresenter(_viewElementFactoryService, textView, parameters);
